DOC: Documentation, plotting, sphinx-gallery fixes (#174)
* DOC: Add skimage as a RTD requirement

* DOC: Reformat README

* FIX: numpy array containing float - invalid index (#169)

* ENH: replace list comprehension with ndarray astype

* correcting origin in plot (within base.py) from "bottomleft" to "lower" (#172)

* FIX: Docstrings, numpy and matplotlib functions

* DOC: Autodoc fixes

+++ doc/_gallery/plot_4_2_2_morlet_scalogram_atoms.py
@@ -47,7 +47,7 @@ axTime.set_xlim(0, 128)
 axTime.set_ylabel('Real part')
 axTime.set_title('Signal in time')
 axTime.grid(True)
-freq_y = np.linspace(0, 0.5, sig.shape[0] / 2)
+freq_y = np.linspace(0, 0.5, int(sig.shape[0] / 2))
 freq_x = (abs(np.fft.fftshift(np.fft.fft(sig))) ** 2)[::-1][:64]
 ix = np.logical_and(freq_y >= 0.05, freq_y <= 0.45)
 axFreq.plot(freq_x[ix], freq_y[ix])
